+++ PublishWorkflow/PublishWorkflowInterface.php
@@ -3,42 +3,20 @@
 namespace Symfony\Cmf\Bundle\CoreBundle\PublishWorkflow;
 
 /**
- * Interface models can implement if they want to support publish workflow checking
+ * Interface models can implement to expose editable publishing settings.
  */
-interface PublishWorkflowInterface
+interface PublishWorkflowInterface extends PublishableInterface, PublishTimePeriodInterface
 {
-    /**
-     * Return the date from which the content should
-     * be considered publishable.
-     *
-     * A NULL value should be interpreted as saying that 
-     * the content has always been publishable.
-     *
-     * @return \DateTime|null
-     */
-    public function getPublishStartDate();
-
     /**
      * Set the date from which the content should
      * be considered publishable.
      *
      * Setting a NULL value asserts that the content
-     * has always been publishable. 
+     * has always been publishable.
      *
      * @param \DateTime|null $publishDate
      */
     public function setPublishStartDate(\DateTime $publishDate = null);
-
-    /**
-     * Return the date at which the content should
-     * stop being published.
-     *
-     * A NULL value should be interpreted as saying that
-     * the document will always be publishable.
-     *
-     * @return \DateTime|null
-     */
-    public function getPublishEndDate();
 
     /**
      * Set the date at which the content should
@@ -52,15 +30,9 @@ interface PublishWorkflowInterface
     public function setPublishEndDate(\DateTime $publishDate = null);
 
     /**
-     * Return the base publishable state of the content.
-     *
-     * A false value indicates that the content should, under
-     * no circumstances, be published. A true value indicates
-     * that the content is publishable - but is subject to
-     * the $publishStartDate and $publishEndDate if they are
-     * set.
+     * Set the boolean flag if this content should be published or not.
      *
      * @return boolean
      */
-    public function isPublishable();
+    public function setPublishable($publishable);
 }